cleaning up API_URLs folder, addressing cant find module
