+++ frontend/app/plaid/page.js
@@ -1,7 +1,7 @@
 'use client'
 import { useState, useEffect, useCallback } from 'react'
 import { usePlaidLink } from 'react-plaid-link'
-import { EXCHANGE_PLAID_TOKEN_URL, CREATE_PLAID_TOKEN_URL } from '@/lib/API_URLS/PROXY'
+import { EXCHANGE_PLAID_TOKEN_URL, CREATE_PLAID_TOKEN_URL } from '@/lib/API_URLS'
 import { Button } from '@/components/ui/button'
 
 export default function PlaidTest() {
